Ensure all standard Ubuntu repos are available
libapache2-mod-fastcgi is provided by Ubuntu's multiverse repository,
which is not enabled by default in Ubuntu Server.

Add an apt_source class which ensures that all of the standard Ubuntu
repositories are available and enabled.

This uses the Puppet apt module, which must be installed before applying
the Puppet config.